refactor: move PDMatrixBuilder class to "style3d/builder.py"
